Fix MacOS crash on startup due to wrong OpenGL profile
The preprocessor directive choose the OpenGL profile in Main.cpp was a condition based on defined(TARGET_OS_IPHONE), which evluated to 1 as the MacOS SDK
had it defined TARGET_OS_IOS to value 0. This caused the OpenGL profile of ES to be used instead of CORE.

Cubic would crash on start up with the following error message:

   SDL_CreateWindow failed: Could not initialize OpenGL / GLES library

Updating the preprocessor to #if TARGET_OS_IOS has the condition evaluate to false when the value is set to zero, which fixes the #if directive to choose the CORE profile intead of ES and make the game run.

Signed-off-by: vlad <vlad@vldr.org>